+++ packages/cheqd-blockchain-modules/src/accumulator/internal.js
@@ -315,12 +315,16 @@ export default class CheqdInternalAccumulatorModule extends injectParams(
       const { additions, removals, witnessUpdateInfo } = CheqdStoredAccumulator.from(
         validateResource(accumulator, String(name), Type),
       );
+      if (witnessUpdateInfo == null) {
+        // eslint-disable-next-line no-continue
+        continue;
+      }
 
       witness.updateUsingPublicInfoPostBatchUpdate(
         member,
         additions ? [...additions].map((addition) => addition.bytes) : [],
         removals ? [...removals].map((removal) => removal.bytes) : [],
-        new VBWitnessUpdateInfo(witnessUpdateInfo?.bytes || new Uint8Array()),
+        new VBWitnessUpdateInfo(witnessUpdateInfo.bytes),
       );
     }
   }