+++ packages/dock-blockchain-modules/src/accumulator/internal.js
@@ -254,15 +254,8 @@ export default class DockInternalAccumulatorModule extends injectParams(
       false,
     );
 
-    let { api } = this.apiProvider;
-    if (!api.registry.hasType('UpdateAccumulator')) {
-      api = await api.at(
-        '0x2e5ae7fdf8d17ddf37de1012d055ac8440dbd2ea57d0dd5a7af5b6dce2e485d6',
-      );
-    }
-
     return extrinsics
-      .map((e) => this.getUpdatesFromExtrinsic(api, e, accumulatorId))
+      .map((e) => this.getUpdatesFromExtrinsic(this.apiProvider.api, e, accumulatorId))
       .filter(Boolean);
   }
 
@@ -281,16 +274,10 @@ export default class DockInternalAccumulatorModule extends injectParams(
         async (b) => await this.apiProvider.getAllExtrinsicsFromBlock(b, false),
       ),
     );
-    let { api } = this.apiProvider;
-    if (!api.registry.hasType('UpdateAccumulator')) {
-      api = await api.at(
-        '0x2e5ae7fdf8d17ddf37de1012d055ac8440dbd2ea57d0dd5a7af5b6dce2e485d6',
-      );
-    }
 
     return extrinsics
       .flat()
-      .map((e) => this.getUpdatesFromExtrinsic(api, e, accumulatorId))
+      .map((e) => this.getUpdatesFromExtrinsic(this.apiProvider.api, e, accumulatorId))
       .filter(Boolean);
   }