+++ packages/dock-blockchain-modules/CHANGELOG.md
@@ -1,5 +1,27 @@
 # @docknetwork/dock-blockchain-modules
 
+## 0.24.0
+
+### Minor Changes
+
+- Introduce `accumulatorVersions`, add unified `addAccumulator` and `updateAccumulator` methods'
+
+### Patch Changes
+
+- Updated dependencies
+  - @docknetwork/credential-sdk@0.35.0
+
+## 0.23.0
+
+### Minor Changes
+
+- Expose `accumulatorHistory`
+
+### Patch Changes
+
+- Updated dependencies
+  - @docknetwork/credential-sdk@0.34.0
+
 ## 0.22.0
 
 ### Minor Changes